+++ packages/nx/lib/ptree.ml
@@ -83,6 +83,24 @@ let instantiate (type a b) (module U : Traverse) :
       U.iter (fun x -> f x) t
   end)
 
+let leaf (type a b) : (module S with type t = (a, b) Nx_effect.t) =
+  (module struct
+    type t = (a, b) Nx_effect.t
+
+    let map (f : 'p 'q. ('p, 'q) Nx_effect.t -> ('p, 'q) Nx_effect.t) (t : t) :
+        t =
+      f t
+
+    let map2
+        (f :
+          'p 'q.
+          ('p, 'q) Nx_effect.t -> ('p, 'q) Nx_effect.t -> ('p, 'q) Nx_effect.t)
+        (a : t) (b : t) : t =
+      f a b
+
+    let iter (f : 'p 'q. ('p, 'q) Nx_effect.t -> unit) (t : t) : unit = f t
+  end)
+
 let unpack ?(at = "") (type a b) (dt : (a, b) Nx_core.Dtype.t) (p : tensor) :
     (a, b) Nx_effect.t =
   match p with